+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - Unreleased
+    - Added progress monitoring to Gradient Boost w/ early stop
     - Added Spatial and Decision tree interface
     - Mean Shift compatible with Spatial trees
     - K-d Neighbors base spatial tree configurable
@@ -23,6 +24,7 @@
     - Gradient Boost shrinkage bounded between 0 and 1
     - Added random subset without replacement to dataset API
     - Changed order of Gradient Boost hyperparameters
+    - Changed order of MLP hyperparameters
 
 - 0.0.13-beta
     - Added documentation site